list toodles-compatible documents in current and nested directories
