+++ packages/app/stores/todos.ts
@@ -10,6 +10,7 @@ import { useTntApi } from '@thombruce/tnt/composables/tntApi'
 
 export const useTodosStore = defineStore('todos', () => {
   // State
+  const document = ref('todo.txt' as String)
   const list = ref([] as Todo[])
 
   // Getters
@@ -29,21 +30,22 @@ export const useTodosStore = defineStore('todos', () => {
   })
 
   // Actions
-  async function fetchTodos() {
-    const file = await useTntApi().loadFile('todo.txt')
+  async function fetchTodos(doc: string) {
+    document.value = doc
+    const file = await useTntApi().loadFile(document.value)
     if (file) list.value = Todo.fromFile(file)
   }
 
   function addTodo(todo: Todo) {
     list.value.push(new Todo(todo))
-    useTntApi().updateFile('todo.txt', Todo.toFile(list.value))
+    useTntApi().updateFile(document.value, Todo.toFile(list.value))
   }
 
   function updateTodoDescription(id: string, description: string, parent?: string) {
     const todo = find.value(id, parent)
     if (!todo) return
     todo.description = description
-    useTntApi().updateFile('todo.txt', Todo.toFile(list.value))
+    useTntApi().updateFile(document.value, Todo.toFile(list.value))
   }
 
   function toggleTodoDone(id: string, parent?: string) {
@@ -56,28 +58,28 @@ export const useTodosStore = defineStore('todos', () => {
       if (next) list.value.push(next)
       todo.close()
     }
-    useTntApi().updateFile('todo.txt', Todo.toFile(list.value))
+    useTntApi().updateFile(document.value, Todo.toFile(list.value))
   }
 
   function toggleTodoFocus(id: string, parent?: string) {
     const todo = find.value(id, parent)
     if (!todo) return
     todo.toggleFocus()
-    useTntApi().updateFile('todo.txt', Todo.toFile(list.value))
+    useTntApi().updateFile(document.value, Todo.toFile(list.value))
   }
 
   function incrementTodoCount(id: string, parent?: string) {
     const todo = find.value(id, parent)
     if (!todo) return
     todo.incrementCount()
-    useTntApi().updateFile('todo.txt', Todo.toFile(list.value))
+    useTntApi().updateFile(document.value, Todo.toFile(list.value))
   }
 
   function toggleTodoTimer(id: string, parent?: string) {
     const todo = find.value(id, parent)
     if (!todo) return
     todo.toggleTimer()
-    if(!todo.isActive) useTntApi().updateFile('todo.txt', Todo.toFile(list.value))
+    if(!todo.isActive) useTntApi().updateFile(document.value, Todo.toFile(list.value))
   }
 
   function deleteTodo(id: string, parent?: string) {
@@ -90,7 +92,7 @@ export const useTodosStore = defineStore('todos', () => {
       list.value = _reject(list.value, { id })
     }
 
-    useTntApi().updateFile('todo.txt', Todo.toFile(list.value))
+    useTntApi().updateFile(document.value, Todo.toFile(list.value))
   }
 
   return {
